Created base input and output components
