+++ Origami/src/components/inputcomponents/ImageInput/ImageInputPreview.js
@@ -2,21 +2,9 @@ import React, { PropTypes } from "react";
 import Dialog from "material-ui/Dialog";
 import FlatButton from "material-ui/FlatButton";
 import ImageInput from "./ImageInput";
+import InputPreview from "../BaseInputComponent/InputPreview.js";
 
-class ImageInputPreview extends React.Component {
-  constructor(props, context) {
-    super(props, context);
-    this.state = {
-      labels: props.functions.getLabels(),
-      open: true
-    };
-    this.hidePreviewDialog = props.functions.hidePreviewDialog;
-    this.handleOk = this.handleOk.bind(this);
-  }
-
-  handleOk() {
-    this.hidePreviewDialog();
-  }
+class ImageInputPreview extends InputPreview {
 
   render() {
     const actions = [
