+++ Origami/src/components/outputcomponents/BaseOutputComponent/OverloadedHeader.js
@@ -2,7 +2,7 @@ import React, { PropTypes } from "react";
 import TextField from "material-ui/TextField";
 import RaisedButton from "material-ui/RaisedButton";
 
-class OverloadedAreaGraphHeader extends React.Component {
+class OverloadedHeader extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -41,8 +41,8 @@ class OverloadedAreaGraphHeader extends React.Component {
   }
 }
 
-OverloadedAreaGraphHeader.propTypes = {
+OverloadedHeader.propTypes = {
   data: PropTypes.object.isRequired
 };
 
-export default OverloadedAreaGraphHeader;
+export default OverloadedHeader;
