+++ Origami/src/components/outputcomponents/ScatterGraphOutput/ScatterGraphOutputPreview.js
@@ -2,21 +2,9 @@ import React, { PropTypes } from "react";
 import Dialog from "material-ui/Dialog";
 import FlatButton from "material-ui/FlatButton";
 import ScatterGraphOutput from "./ScatterGraphOutput";
+import OutputPreview from "../BaseOutputComponent/OutputPreview.js";
 
-class ScatterGraphOutputPreview extends React.Component {
-  constructor(props, context) {
-    super(props, context);
-    this.state = {
-      headers: props.functions.getHeaders(),
-      open: true
-    };
-    this.hidePreviewDialog = props.functions.hidePreviewDialog;
-    this.handleOk = this.handleOk.bind(this);
-  }
-
-  handleOk() {
-    this.hidePreviewDialog();
-  }
+class ScatterGraphOutputPreview extends OutputPreview {
 
   render() {
     const actions = [
