+++ Origami/src/components/outputcomponents/TextOutput/TextOutputPreview.js
@@ -2,22 +2,10 @@ import React, { PropTypes } from "react";
 import Dialog from "material-ui/Dialog";
 import FlatButton from "material-ui/FlatButton";
 import TextOutput from "./TextOutput";
+import OutputPreview from "../BaseOutputComponent/OutputPreview.js";
 
-class TextOutputPreview extends React.Component {
-  constructor(props, context) {
-    super(props, context);
-    this.state = {
-      headers: props.functions.getHeaders(),
-      open: true
-    };
-    this.hidePreviewDialog = props.functions.hidePreviewDialog;
-    this.handleOk = this.handleOk.bind(this);
-  }
-
-  handleOk() {
-    this.hidePreviewDialog();
-  }
-
+class TextOutputPreview extends OutputPreview {
+  
   render() {
     const actions = [
       <FlatButton
